PHP 게시판 수정
- 답글 정렬 이상했던 수정.
- 댓글 갯수 목록에 표시
- 첨부파일 갯수 표시
- 관리자가 게시글 삭제할 때 첨부파일 삭제되게 수정.
- 비회원인데 수정 안되는것 수정

+++ board/delete.php
@@ -1,17 +1,15 @@
 <?php
     require_once("../common.php");
 
-    extract($_GET);
-
     if (!isset($bno) || !$bno) {
         alert("잘못된 접근입니다.");
         move();
     }
 
-    if ($_SESSION['user_authority'] == "ADMIN") {
+    if (isset($_SESSION['user_authority']) && $_SESSION['user_authority'] == "ADMIN") {
         $sql = "DELETE FROM board WHERE bno = ?";
         $params = [$bno];
-    } else if ( $_SESSION['user_authority'] == "MEMBER" ) {
+    } else if ( isset($_SESSION['user_authority']) && $_SESSION['user_authority'] == "MEMBER" ) {
         $sql = "DELETE FROM board WHERE bno = ? AND id = ?";
         $params = [$bno, $_SESSION['user_id']];
     } else {
@@ -25,7 +23,7 @@
             move("list.php");
         } else {
             alert("잘못된 접근입니다.");
-            move();
+            //move();
         }
     } catch (Exception $e) {
         alert("글수정을 실패하셨습니다.\\n시스템 관리자에게 문의하세요.");
